Rebuild OHW22 toctree
It looks like a page needs to be in an existing toctree, before Sphinx will build a toctree for it.

I've added it back to the primary toctree, but I've limited the number of entries that will display in the menu.

I've also gone and removed the duplicate search options from the menu while I was there.

Closes #245